requirements.txt
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
Django==5.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
python-telegram-bot==21.4
sniffio==1.3.1
sqlparse==0.5.1
tzdata==2024.1
